Added the possibility to add players from the FIDE and FFE databases (#123)
* Fixed FR translation

* Added PapiWebConfig.federation_database_ext
Simplified FFEDatabase and FideDatabase

* Fixed FR translation

* Added the possibility to add players from the FFE database

* Fixed the number of results returned by an FFE search

* Fixed the players' rating when adding a player

* Fixed typo in /admin/players/tab.html

* Display the FFE licence number when expired

* Added FIDE search (not working yet)

* Removed debug

* Fixed FR translation

* Added a hidden value for disabled text inputs

* Added all the available information from the FIDE and FFE databases when adding a player

* Finalized the entry of new players in tournaments

* Updated the translations

* Fixed player deletion

* Fixed FIDE and FFE downloads

* Fixed the creation of players with FFE ID but no FIDE ID

* Escape the search tokens passed to the FIDE and FFE searches

* Added a dependency to dateutil

* Sanitized the search tokens

* Fixed the dependency to dateutil

* Removed useless pass

* Sanitized limit

+++ src/database/sqlite/fide_database.py
@@ -1,4 +1,5 @@
 import os.path
+from string import capwords
 from xml.etree import ElementTree
 import zipfile
 from contextlib import suppress
@@ -47,8 +48,8 @@ class FideDatabase(SQLiteDatabase):
             ...
     """
 
-    def __init__(self, basename: str = 'fide', write: bool = False):
-        super().__init__(TMP_DIR / f'{basename}.{PapiWebConfig.event_ext}', write)
+    def __init__(self, write: bool = False):
+        super().__init__(TMP_DIR / f'fide.{PapiWebConfig.federation_database_ext}', write)
 
     def check(self) -> bool:
         """Check if the database exists and proposes to create it if not, or update it if too old,
@@ -159,9 +160,6 @@ class FideDatabase(SQLiteDatabase):
                         if field_function:
                             data[field_name] = field_function(data[field_name])
                     if ',' in data['name']:
-                        name_parts: list[str] = data['name'].split(',', maxsplit=2)
-                        data['last_name'] = name_parts[0]
-                        data['first_name'] = name_parts[1]
                         data['last_name'], data['first_name'] = data['name'].split(
                             ',', maxsplit=1
                         )
@@ -181,7 +179,8 @@ class FideDatabase(SQLiteDatabase):
             if save:
                 save.rename(self.file)
             return False
-        save.unlink(missing_ok=True)
+        if save:
+            save.unlink(missing_ok=True)
         print_interactive_success(
             _('{number} players written.').format(number=players_number)
         )
@@ -194,7 +193,54 @@ class FideDatabase(SQLiteDatabase):
         )
         yield from map(lambda row: row['federation'], self._fetchall())
 
-    def search_player(self, string: str) -> Iterator[Player]:
+    @staticmethod
+    def get_player_from_row(row: dict[str, Any]) -> Player | None:
+        return Player(
+            id=0,
+            first_name=capwords(row['first_name']),
+            last_name=row['last_name'].upper(),
+            date_of_birth=datetime.strptime(
+                f'{row["year_of_birth"] or 1900}-01-01', '%Y-%m-%d'
+            ).date(),
+            gender=PlayerGender(row['gender']),
+            mail='',
+            phone='',
+            comment='',
+            owed=0.0,
+            paid=0.0,
+            title=PlayerTitle(row['fide_title']),
+            ratings={
+                TournamentRating.STANDARD: row['standard_rating'],
+                TournamentRating.RAPID: row['rapid_rating'],
+                TournamentRating.BLITZ: row['blitz_rating'],
+            },
+            rating_types={
+                TournamentRating.STANDARD:
+                    PlayerRatingType.FIDE if row['standard_rating'] else PlayerRatingType.ESTIMATED,
+                TournamentRating.RAPID:
+                    PlayerRatingType.FIDE if row['rapid_rating'] else PlayerRatingType.ESTIMATED,
+                TournamentRating.BLITZ:
+                    PlayerRatingType.FIDE if row['blitz_rating'] else PlayerRatingType.ESTIMATED,
+            },
+            fide_id=row['fide_id'],
+            ffe_id=None,
+            ffe_licence=PlayerFFELicence.NONE,
+            ffe_licence_number=None,
+            federation=row['federation'],
+            league='',
+            club='',
+            fixed=0,
+            check_in=False,  # not taken into account when updating/creating/deleting the player
+            pairings={},  # Pairings are read from Papi but not used
+            tournament=None,
+        ) if row else None
+
+
+    def search_player(
+            self,
+            string: str,
+            limit: int = 0,  # no limit set if no param or null param passed
+    ) -> Iterator[Player]:
         tokens: list[str] = string.split(' ')
         str_fields: tuple[str, ...] = (
             'last_name',
@@ -202,65 +248,34 @@ class FideDatabase(SQLiteDatabase):
         )
         int_fields: tuple[str, ...] = ('fide_id',)
         token_conditions: dict[str, str] = {}
+        params: list[Any] = []
         for token in tokens:
             expressions = list(
-                map(lambda field: f"({field} LIKE '%{token}%')", str_fields)
+                map(lambda field: f'({field} LIKE ?)', str_fields)
             )
+            params += [f'%{token}%', ] * len(str_fields)
             int_value: int
             with suppress(ValueError):
                 int_value = int(token.strip())
                 expressions += list(
-                    map(lambda field: f'({field} = {int_value})', int_fields)
+                    map(lambda field: f'({field} = ?)', int_fields)
                 )
+                params += [f'%{int_value}%', ] * len(int_fields)
             token_conditions[token] = ' OR '.join(expressions)
         conditions: str = ' AND '.join(
             map(lambda condition: f'({condition})', token_conditions.values())
         )
-        self._execute(f'SELECT * FROM player WHERE {conditions}')
+        query: str = f'SELECT * FROM player WHERE {conditions}'
+        if limit:
+            query += ' LIMIT ?'
+            params += [limit, ]
+        self._execute(query, tuple(params), )
         return (
-            Player(
-                id=0,
-                first_name=row['first_name'],
-                last_name=row['last_name'],
-                date_of_birth=datetime.strptime(
-                    f'{row["year_of_birth"]}-01-01', '%Y-%m-%d'
-                ).date()
-                if row['year_of_birth']
-                else None,
-                gender=PlayerGender(row['gender']),
-                mail='',
-                phone='',
-                comment='',
-                owed=0.0,
-                paid=0.0,
-                title=PlayerTitle(row['fide_title']),
-                ratings={
-                    TournamentRating.STANDARD: row['standard_rating'],
-                    TournamentRating.RAPID: row['rapid_rating'],
-                    TournamentRating.BLITZ: row['blitz_rating'],
-                },
-                rating_types={
-                    TournamentRating.STANDARD: PlayerRatingType.FIDE
-                    if row['standard_rating']
-                    else PlayerRatingType.ESTIMATED,
-                    TournamentRating.RAPID: PlayerRatingType.FIDE
-                    if row['rapid_rating']
-                    else PlayerRatingType.ESTIMATED,
-                    TournamentRating.BLITZ: PlayerRatingType.FIDE
-                    if row['blitz_rating']
-                    else PlayerRatingType.ESTIMATED,
-                },
-                fide_id=row['fide_id'],
-                ffe_id=0,
-                ffe_licence=PlayerFFELicence.NONE,
-                ffe_licence_number=None,
-                federation=row['federation'],
-                league='',
-                club='',
-                fixed=0,
-                check_in=False,  # not taken into account when updating/creating/deleting the player
-                pairings={},  # Pairings are read from Papi but not used
-                tournament=None,
-            )
+            self.get_player_from_row(row)
             for row in self._fetchall()
         )
+
+
+    def get_player_by_fide_id(self, player_fide_id: int) -> Player | None:
+        self._execute(f'SELECT * FROM player WHERE fide_id = ?', (player_fide_id, ))
+        return self.get_player_from_row(self._fetchone())
